Experts for the Point Environment
The experts are trained using the PPO algorithm in the script
train_point_experts, and the corresponding trained networks are dumped
in tensorflow sessions in the folder ppo_point_expert.
Those sessions are loaded in the script sample_point_expert where
rollouts are ran to sample trajectories from the expert.
Those trajectories are then saved in numpy files under the folder
point_data.