Create @wrap_experiment to replace run_experiment
This implements a minimal version of the interface described in #1049.
It has a few limitations and doesn't implement every feature we might
need, but it's already probably better than run_experiment.

I believe this implementation can be cleanly extended to provide
everything we want from the interface.

The most important difference (besides all the new features) is that it
doesn't call garage.deterministic.set_seed. Since this is a parameter
that should be logged in variant.json, it should actually be passed into
the experiment function. Perhaps the runner should call this function
instead?

This commit does not contain tests or examples, only the implementation
and tests.