Add method for adding samples to PathBuffer
This allows adding samples without requiring them to be in a path.
Although these samples can't ever be sampled using sample_path, they can
still be sampled with sample_transitions.

This is useful if samples arrive in complex sequences, and the method
sampling from the buffer doesn't want to sample paths.